@@ -7,7 +7,7 @@
     <groupId>com.github.nutt1101</groupId>
     <artifactId>CatchBall</artifactId>
 
-    <version>1.1.6</version>
+    <version>1.1.7</version>
 
     <modules>
         <module>v1_16</module>
@@ -22,7 +22,7 @@
     <name>CatchBall</name>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>16</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
